refactor: extract file-upload.ts to make it reusable
