--- backend/src/items/file-upload/storage-options.ts
@@ -1,11 +1,0 @@
-import { diskStorage } from 'multer';
-import { editFileName, validateFileType } from './file.utils';
-
-export const storageOptions = {
-  storage: diskStorage({
-    destination: './src/uploads/item-images',
-    filename: editFileName,
-  }),
-  fileFilter: validateFileType,
-  limits: { fileSize: 5 * 1024 * 1024 },
-};
